Add QmlPromise class
This class is useful for creating promises in C++ which are then passed
to QML. QML is then able to call .then() on these promises to add
continuations. The C++ QmlPromise can then be used to resolve or reject
the promise, thus allowing QML to be notified when a C++ asynchronous
operation finishes.